Pyrolite trunk checkin

+++ build-dotnet-mono.sh
@@ -1,9 +1,11 @@
 #!/bin/sh
 echo "Compiling .net source"
-rm -r dotnet/Pyrolite/bin
-#rm -r dotnet/Pyrolite/obj
-rm -r dotnet/Pyrolite.Tests/bin
-#rm -r dotnet/Pyrolite.Tests/obj
-xbuild /verbosity:minimal /property:Configuration=Debug dotnet/Pyrolite.sln
+if [ -d dotnet/Pyrolite/bin ]; then
+  rm -r dotnet/Pyrolite/bin
+fi
+if [ -d dotnet/Pyrolite.Tests/bin ]; then
+  rm -r dotnet/Pyrolite.Tests/bin
+fi
+xbuild /verbosity:minimal /property:Configuration=Debug /property:Platform="Any CPU" dotnet/Pyrolite.sln
 cp dotnet/Pyrolite/bin/Debug/*.dll build
